requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
gspread>=6.1.0
google-auth>=2.29.0
python-dotenv>=1.0.1
Pillow>=10.3.0
